WIP merge the GL into to SDL to gain better control on the rendering
Next need to refactor the core to get rid of the callback traits and just return a ref to the buffer after `cycle_frame`
