+++ common/src/initialization.rs
@@ -1,10 +1,10 @@
 use log::info;
 
-use magenboy_core::{AudioDevice, Bootrom, GameBoy, JoypadProvider, Mode, GBC_BOOT_ROM_SIZE, GB_BOOT_ROM_SIZE};
+use magenboy_core::{mmu::carts::Mbc, AudioDevice, Bootrom, GameBoy, GfxDevice, JoypadProvider, Mode, GBC_BOOT_ROM_SIZE, GB_BOOT_ROM_SIZE};
 #[cfg(feature = "dbg")]
 use magenboy_core::debugger::DebuggerInterface;
 
-use crate::{mbc_handler::{initialize_mbc, release_mbc}, menu::MagenBoyState, mpmc_gfx_device::MpmcGfxDevice};        
+use crate::menu::MagenBoyState;
 
 pub fn check_for_terminal_feature_flag(args:&Vec::<String>, flag:&str)->bool{
     args.len() >= 3 && args.contains(&String::from(flag))
@@ -18,14 +18,14 @@ pub fn get_terminal_feature_flag_value(args:&Vec<String>, flag:&str, error_messa
 // This is static and not local for the unix signal handler to access it
 pub static EMULATOR_STATE:MagenBoyState = MagenBoyState::new();
 
-pub fn init_and_run_gameboy(
+pub fn init_gameboy<'a>(
     args: Vec<String>,
-    program_name: String, 
-    spsc_gfx_device: MpmcGfxDevice, 
+    mbc: &'a mut dyn Mbc, 
+    gfx_device: impl GfxDevice, 
     joypad_provider: impl JoypadProvider,
     audio_devices: impl AudioDevice,
     #[cfg(feature = "dbg")] dui: impl DebuggerInterface
-){
+)-> GameBoy<'a, impl JoypadProvider, impl AudioDevice, impl GfxDevice>{
     let bootrom_path = if check_for_terminal_feature_flag(&args, "--bootrom"){
         Some(get_terminal_feature_flag_value(&args, "--bootrom", "Error! you must specify a value for the --bootrom parameter"))
     }else{
@@ -34,10 +34,8 @@ pub fn init_and_run_gameboy(
 
     let bootrom = read_bootrom(bootrom_path);
 
-    let mbc = initialize_mbc(&program_name);
-
-    let mut gameboy = match bootrom{
-        Some(b) => GameBoy::new_with_bootrom(mbc, joypad_provider, audio_devices, spsc_gfx_device, b, #[cfg(feature = "dbg")] dui),
+    let gameboy = match bootrom{
+        Some(b) => GameBoy::new_with_bootrom(mbc, joypad_provider, audio_devices, gfx_device, b, #[cfg(feature = "dbg")] dui),
         None => {
             let mode = if check_for_terminal_feature_flag(&args, "--mode"){
                 let mode = get_terminal_feature_flag_value(&args, "--mode", "Error: Must specify a mode");
@@ -49,24 +47,15 @@ pub fn init_and_run_gameboy(
                 log::info!("Could not find a mode flag, auto detected {}", <Mode as Into<&str>>::into(mode));
                 mode
             };
-            GameBoy::new_with_mode(mbc, joypad_provider, audio_devices, spsc_gfx_device, mode, #[cfg(feature = "dbg")] dui)
+            GameBoy::new_with_mode(mbc, joypad_provider, audio_devices, gfx_device, mode, #[cfg(feature = "dbg")] dui)
         }
     };
 
     info!("initialized gameboy successfully!");
 
     EMULATOR_STATE.running.store(true, std::sync::atomic::Ordering::Relaxed);
-    while EMULATOR_STATE.running.load(std::sync::atomic::Ordering::Relaxed){
-        if !EMULATOR_STATE.pause.load(std::sync::atomic::Ordering::SeqCst){
-            // Locking the state mutex in order to signal the menu that we are cycling a frame now
-            let state = &EMULATOR_STATE;
-            let _mutex_ctx = state.state_mutex.lock().unwrap();
-            gameboy.cycle_frame();
-        }
-    }
-    drop(gameboy);
-    release_mbc(&program_name, mbc);
-    log::info!("released the gameboy succefully");
+
+    return gameboy;
 }
 
 pub fn read_bootrom(bootrom_path: Option<String>) -> Option<Bootrom> {
