+++ common/src/menu.rs
@@ -19,21 +19,19 @@ pub const GAME_MENU_OPTIONS:[MenuOption<EmulatorMenuOption, &str>;3] = [
 ];
 
 cfg_if::cfg_if!{ if #[cfg(feature = "std")]{
-    use std::{sync::{atomic::AtomicBool, Mutex}, path::PathBuf};
+    use std::{sync::atomic::AtomicBool, path::PathBuf};
     use magenboy_core::{ppu::gfx_device::GfxDevice, keypad::joypad_provider::JoypadProvider};
     use super::joypad_menu::{MenuJoypadProvider, joypad_gfx_menu, JoypadMenu, MenuRenderer};
 
     pub struct MagenBoyState{
         // Use atomic bool, normal bool doesnt works on arm (probably cause of the memory model)
         pub running:AtomicBool,
-        pub pause:AtomicBool,
         pub exit:AtomicBool,
-        pub state_mutex:Mutex<()>
     }
 
     impl MagenBoyState{
         pub const fn new() -> Self {
-            Self { running: AtomicBool::new(true), pause: AtomicBool::new(false), exit: AtomicBool::new(false), state_mutex: Mutex::new(()) }
+            Self { running: AtomicBool::new(true), exit: AtomicBool::new(false) }
         }
     }
 
@@ -47,8 +45,8 @@ cfg_if::cfg_if!{ if #[cfg(feature = "std")]{
             Self { provider, header }
         }
 
-        pub fn pop_game_menu<GFX:GfxDevice>(&mut self, state:&MagenBoyState, gfx_device:&mut GFX, receiver:crossbeam_channel::Receiver<usize>){
-            match self.get_game_menu_selection(state, gfx_device, receiver){
+        pub fn pop_game_menu<GFX:GfxDevice>(&mut self, state:&MagenBoyState, gfx_device:&mut GFX){
+            match self.get_game_menu_selection(gfx_device){
                 EmulatorMenuOption::Resume => {},
                 EmulatorMenuOption::Restart => state.running.store(false, std::sync::atomic::Ordering::Relaxed),
                 EmulatorMenuOption::Shutdown => {
@@ -58,24 +56,10 @@ cfg_if::cfg_if!{ if #[cfg(feature = "std")]{
             }
         }
 
-        fn get_game_menu_selection<GFX:GfxDevice>(&mut self, state:&MagenBoyState,gfx_device:&mut GFX, emulation_framebuffer_channel:crossbeam_channel::Receiver<usize>)->&EmulatorMenuOption{
+        fn get_game_menu_selection<GFX:GfxDevice>(&mut self, gfx_device:&mut GFX)->&EmulatorMenuOption{
             let menu_renderer = joypad_gfx_menu::GfxDeviceMenuRenderer::new(gfx_device);
-        
             let mut menu = JoypadMenu::new(&GAME_MENU_OPTIONS, &self.header, menu_renderer);  
-        
-            // lock the mutex here to sync the 2 threads
-            state.pause.store(true, std::sync::atomic::Ordering::SeqCst);
-            loop{
-                if let Ok(_lock) = state.state_mutex.try_lock(){
-                    let selection = menu.get_menu_selection(&mut self.provider);
-                    state.pause.store(false, std::sync::atomic::Ordering::SeqCst);
-                    return selection;
-                }else{
-                    // try recv in order to clear frames from the channel 
-                    // in order to not block the emualtion thread and allow it to finish the frame
-                    let _ = emulation_framebuffer_channel.try_recv();
-                }
-            }
+            return menu.get_menu_selection(&mut self.provider);
         }
 
     }
