--- docs/GL.md
@@ -1,5 +1,0 @@
-# GL (OpenGL POC)
-
-This is a POC frontend for GLFW + OpenGL frontend in Rust
-
-Currently no support for audio.
