--- gl/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "magenboy_gl"
-version.workspace = true
-authors.workspace = true
-rust-version.workspace = true
-edition.workspace = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-magenboy_core ={ path = "../core" }
-magenboy_common = { path = "../common", features = ["std"] }
-glfw-sys ={ version="6.0", default-features = false, features = ["src-build", "static-link"] }
-gl = "0.14"
-argh = "0.1.13"
