+++ sdl/Cargo.toml
@@ -16,10 +16,11 @@ path = "src/main.rs"
 magenboy_core = {path = "../core/", features = ["apu"]}
 magenboy_common = {path = "../common/", features = ["std"]}
 log = {version = "0.4", features = ["max_level_debug", "release_max_level_info"]}
-sdl2 = "0.35"
+sdl2 = "0.38"
 crossbeam-channel = "0.5"
 cfg-if = "1.0"
 wav = "1.0"
+gl = "0.14"
 
 [features]
 default = ["static-sdl"]
