+++ sdl/build.rs
@@ -1,0 +1,4 @@
+fn main() {
+    #[cfg(target_os="linux")]
+    println!("cargo:rustc-link-arg=-Wl,-rpath,$ORIGIN");
+}
