+++ sdl/src/audio/sdl_audio_device.rs
@@ -64,7 +64,6 @@ impl<AR:AudioResampler> ResampledAudioDevice<AR> for SdlAudioDevice<AR>{
         };
 
         unsafe{
-            SDL_Init(SDL_INIT_AUDIO);
             SDL_ClearError();
             let mut uninit_audio_spec:MaybeUninit<SDL_AudioSpec> = MaybeUninit::uninit();
             let id = SDL_OpenAudioDevice(std::ptr::null(), 0, &desired_audio_spec, uninit_audio_spec.as_mut_ptr() , 0);
