+++ sdl/src/utils.rs
@@ -3,9 +3,9 @@ use std::ffi::CStr;
 use sdl2::{libc::c_char, sys::*};
 
 pub fn get_sdl_error_message()->&'static str{
+    // SAFETY: returned str from SDL_GetError is static and therfore can be used with static lifetime
     unsafe{
         let error_message:*const c_char = SDL_GetError();
-        
         return CStr::from_ptr(error_message).to_str().unwrap();
     }
 }
